wdio-conifig: move validateConfig to webdriver package
