+++ packages/webdriver/package.json
@@ -30,7 +30,6 @@
   "dependencies": {
     "deepmerge": "^2.0.1",
     "request": "^2.83.0",
-    "wdio-config": "^5.0.0-beta.3",
     "wdio-logger": "^5.0.0-beta.3"
   }
 }
